Add files via upload
The number of goals is now based on the value of G, which is now determined by the length of V, which can be altered in the JS file for different results. It also shows both players' progress towards these goals, which can be changed in the array S - also in the JS file. Currently the program assumes that all goals are open and have the same value to both players.
The bug that prevented the scripts in the JS document from interacting with the HTML document has been fixed.